# api/swagger/marketplace.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Marketplace API",
    "version": "1.0"
  },
  "host": "marketplace.dev.videocoin.network",
  "schemes": [
    "https"
  ],
  "consumes": [
    "application/json",
    "multipart/form-data"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/v1/account": {
      "get": {
        "summary": "Get account info",
        "operationId": "GetAccount",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/AccountResponse"
            }
          },
          "401": {
            "description": "Returned when the user is not logged in.",
            "schema": {}
          },
          "403": {
            "description": "Returned when the user does not have permission to access the resource.",
            "schema": {}
          },
          "404": {
            "description": "Returned when the resource does not exist.",
            "schema": {
              "example": {
                "message": "Not found",
                "fields": null
              }
            }
          }
        },
        "tags": [
          "Proifle"
        ]
      },
      "put": {
        "summary": "Update account info",
        "operationId": "UpdateAccount",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/AccountResponse"
            }
          },
          "401": {
            "description": "Returned when the user is not logged in.",
            "schema": {}
          },
          "403": {
            "description": "Returned when the user does not have permission to access the resource.",
            "schema": {}
          },
          "404": {
            "description": "Returned when the resource does not exist.",
            "schema": {
              "example": {
                "message": "Not found",
                "fields": null
              }
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateAccountRequest"
            }
          }
        ],
        "tags": [
          "Proifle"
        ]
      }
    },
...
    "/api/v1/accounts/{address}/nonce": {
      "get": {
        "summary": "Get nonce by address account",
        "operationId": "GetNonce",
        "security": [],
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/NonceResponse"
            }
          },
          "401": {
            "description": "Returned when the user is not logged in.",
            "schema": {}
          },
          "403": {
            "description": "Returned when the user does not have permission to access the resource.",
            "schema": {}
          },
          "404": {
            "description": "Returned when the resource does not exist.",
            "schema": {
              "example": {
                "message": "Not found",
                "fields": null
              }
            }
          }
        },
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "Registration"
        ]
      }
    },
...
  },
  "definitions": {
...
    "NonceResponse": {
      "type": "object",
      "properties": {
        "nonce": {
          "type": "string"
        }
      }
    },
    "UpdateAccountRequest": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "bio": {
          "type": "string"
        },
        "custom_url": {
          "type": "string"
        },
        "yt_username": {
          "type": "string"
        },
        "image_data": {
          "type": "string"
        },
        "cover_data": {
          "type": "string"
        }
      }
    },
...
    "UserResponse": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "cover_url": {
          "type": "string"
        },
        "bio": {
          "type": "string"
        },
        "custom_url": {
          "type": "string"
        },
        "yt_username": {
          "type": "string"
        }
      }
    },
    "AccountResponse": {
      "type": "object",
      "properties": {
        "address": {
          "type": "string"
        },
        "profile_img_url": {
          "type": "string"
        },
        "is_verified": {
          "type": "boolean"
        },
        "user": {
          "$ref": "#/definitions/UserResponse"
        }
      }
    },
...
  },
  "securityDefinitions": {
    "ApiKeyAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "security": [
    {
      "ApiKeyAuth": []
    }
  ]
}